security(wave-4): process hardening + CI gates + module split + docs
Wave 4 of the 5-wave plan. Process and structure work: prevent the
audit findings from regressing, document the trust model, and pay down
the technical debt that the Wave 3 server-only push exposed.

Module split (closes the Wave 3 escape hatch)
- New src/lib/plan-shared.ts: all plan types + pure helpers (slug,
  label, summarize, group). plan.ts re-exports them and keeps only the
  Supabase DB calls. Adds `import "server-only"` to plan.ts so any
  future client import fails the build instead of silently bundling
  the Supabase SDK.
- New src/lib/hall-compose-shared.ts: HallDraft types + withDraftDefaults
  helper. hall-compose.ts re-exports them and gains `import "server-only"`.
- Client components rewired: PlanView, ArtifactRow, CreateDraftPanel,
  HallDraftReviewClient, HallComposedHero now import from -shared.
- src/lib/supabase.ts gets `import "server-only"` (the Wave 3 escape
  hatch that pulled this when plan.ts imported it is now closed).

Logging hygiene sweep (audit M7)
- Apply the "log server-side, return generic" pattern to the routes the
  audit specifically flagged for echoing err.message / String(err):
  /api/upload, /api/garage-ingest (file download path),
  /api/garage-upload (POST + DELETE), /api/projects/[id]/proposal-upload
  (storage, signed URL, db update branches). Full sweep across the
  remaining ~20 routes (mostly admin-only) tracked for follow-up.

ESLint: react/jsx-no-target-blank (audit LOW)
- Added as `error` with `allowReferrer: false`. Verified: 78
  target="_blank" occurrences scanned, all already carry rel.

CI gates (audit Wave 4 plan)
- scripts/check-secrets.sh — pattern scan for known leaked secret
  shapes (Notion, Clerk, Anthropic, Google, JWTs, PEMs, historical
  literals `ch-os-agent-2024-secure` + `ch-agents-2026`). Hook'd via
  .husky/pre-commit on staged diffs.
- scripts/check-api-auth.sh — every mutating /api/**/route.ts must
  reference one of the auth helpers (adminGuardApi, requireCronAuth,
  isValidCronRequest, currentUser, requireAdminAction,
  requireSameOriginRequest) or be on the public allowlist. Verified
  172 routes scanned, all guarded.
- .github/workflows/security-gates.yml — wires both scripts to GitHub
  Actions, plus npm audit gate (fails on net-new high/critical CVEs
  over the documented baseline of 5 transitive d3-color highs) and
  tsc --noEmit. Runs on every PR + push to main.
- .github/dependabot.yml — weekly npm + github-actions updates,
  grouped by minor/patch, ignored majors on Next/React/Clerk to avoid
  breaking auto-merges. Security advisories arrive out-of-band.

docs/SECURITY_MODEL.md
- Canonical statement of identities, trust boundaries, secret blast
  radius table with rotation playbooks per secret, CI gates,
  public-by-design routes, defense-in-depth boundaries, known
  accepted risks, incident response, and audit history (waves 0–4).
