+++ eslint.config.mjs
@@ -5,6 +5,16 @@ import nextTs from "eslint-config-next/typescript";
 const eslintConfig = defineConfig([
   ...nextVitals,
   ...nextTs,
+  // Security: every <a target="_blank"> must carry rel="noopener noreferrer"
+  // to prevent reverse-tabnabbing. The audit found ~80 occurrences.
+  {
+    rules: {
+      "react/jsx-no-target-blank": [
+        "error",
+        { allowReferrer: false, enforceDynamicLinks: "always", warnOnSpreadAttributes: true, links: true, forms: true },
+      ],
+    },
+  },
   // Override default ignores of eslint-config-next.
   globalIgnores([
     // Default ignores of eslint-config-next:
