+++ src/app/api/garage-ingest/route.ts
@@ -291,10 +291,8 @@ export async function POST(req: NextRequest) {
     if (err instanceof SsrfBlockedError) {
       return NextResponse.json({ error: "URL rejected by SSRF policy", detail: err.message }, { status: 400 });
     }
-    return NextResponse.json(
-      { error: `Failed to download file: ${err instanceof Error ? err.message : String(err)}` },
-      { status: 502 }
-    );
+    console.error("[garage-ingest] file download failed:", err);
+    return NextResponse.json({ error: "Failed to download file" }, { status: 502 });
   }
 
   // ── Step 2: Detect file type and build Anthropic message ───────────────────
