+++ src/lib/supabase.ts
@@ -1,10 +1,4 @@
-// NOTE: server-only marker intentionally OMITTED.
-// src/lib/plan.ts imports supabaseAdmin and is in turn imported by client
-// components (PlanView, ArtifactRow, CreateDraftPanel). Adding "server-only"
-// here breaks the build until plan.ts is split into client-safe types vs
-// server DB calls (tracked as Wave 4). Defense-in-depth via server-only
-// is in effect on supabase-server.ts, notion/core.ts, drive.ts, and
-// google-auth.ts.
+import "server-only";
 import { createClient, SupabaseClient } from "@supabase/supabase-js";
 
 let _admin: SupabaseClient | null = null;
